ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
openai==1.3.7
httpx==0.25.2
pydantic==2.5.2
python-dotenv==1.0.0
